Tile to activate the alarm, not the light
The siren tile in Homey UI will now set of the alarm as defined in the settings. The light can now be set if selecting the Siren and switching the Light on. Also the title of the Light is now referring to the Light and not On/Off as this is confusing for a Siren device